+++ package-stubber/community-stubs/uploader.js
@@ -1,0 +1,8 @@
+// uploader package
+
+function emptyFn () {}
+
+Uploader = {
+  config: emptyFn,
+  getConfig: emptyFn
+};